Fix the bug of the top wave breaking off by 1px at certain resolutions
Researched how svg code works enough to implement a gradient, and then
matched the gradient to the one used for the background. Then inverted
the wave direction and moved it to the companies section instead of the
intro section.

In practice, it should not "separate" by one pixel anymore because it's
now an additive effect: rather than a superimposed white effect on top
of a gradient background, it is now a gradient effect superimposed to
the white background below. This hides the transition whenever the reso-
lution moves the wave's box a pixel up or down.

The one con to this approach is that, IF we were to change the theme
colors of the page, the fill of the wave would now have to also be
changed manually. But since that is already the case for other parts of
the page (mainly the pics downloaded from manipixels) the change is all
pros.